requirements.txt
pandas
google-cloud-bigquery
google-cloud-storage
tqdm
firecloud
PyCap
mysql-connector-python
